+++ jest.setup.js
@@ -1,0 +1,1 @@
+process.env._YARN_PLUGIN_AWS_CODEARTIFACT_TESTING = "true";