+++ src/__tests__/integration/authHook.test.ts
@@ -3,7 +3,6 @@ import { join } from "path";
 
 const env = {
   ...process.env,
-  _YARN_PLUGIN_AWS_CODEARTIFACT_TESTING: "true",
   _YARN_PLUGIN_AWS_CODEARTIFACT_DEBUG: "true",
 };
 const expectedRegex = /Setting token for registry .+ to ~~~domain-test~000000000000~us-east-1~aws-profile-2~true~~~/;